+++ internal/component/database_observability/mysql/collector/schema_table.go
@@ -185,7 +185,7 @@ func (c *SchemaTable) Name() string {
 }
 
 func (c *SchemaTable) Start(ctx context.Context) error {
-	level.Debug(c.logger).Log("msg", SchemaTableName+" collector started")
+	level.Debug(c.logger).Log("msg", "collector started")
 
 	c.running.Store(true)
 	ctx, cancel := context.WithCancel(ctx)
@@ -229,8 +229,7 @@ func (c *SchemaTable) Stop() {
 func (c *SchemaTable) extractSchema(ctx context.Context) error {
 	rs, err := c.dbConnection.QueryContext(ctx, selectSchemaName)
 	if err != nil {
-		level.Error(c.logger).Log("msg", "failed to query schemata", "err", err)
-		return err
+		return fmt.Errorf("failed to query schemata: %w", err)
 	}
 	defer rs.Close()
 
@@ -251,8 +250,7 @@ func (c *SchemaTable) extractSchema(ctx context.Context) error {
 	}
 
 	if err := rs.Err(); err != nil {
-		level.Error(c.logger).Log("msg", "error during iterating over schemas result set", "err", err)
-		return err
+		return fmt.Errorf("failed to iterate over schemas result set: %w", err)
 	}
 
 	if len(schemas) == 0 {
@@ -295,8 +293,7 @@ func (c *SchemaTable) extractSchema(ctx context.Context) error {
 		}
 
 		if err := rs.Err(); err != nil {
-			level.Error(c.logger).Log("msg", "error during iterating over tables result set", "err", err)
-			return err
+			return fmt.Errorf("failed to iterate over tables result set: %w", err)
 		}
 	}
 
@@ -421,7 +418,7 @@ func (c *SchemaTable) fetchColumnsDefinitions(ctx context.Context, schemaName st
 	}
 
 	if err := colRS.Err(); err != nil {
-		level.Error(c.logger).Log("msg", "error during iterating over table columns result set", "schema", schemaName, "table", tableName, "err", err)
+		level.Error(c.logger).Log("msg", "failed to iterate over table columns result set", "schema", schemaName, "table", tableName, "err", err)
 		return nil, err
 	}
 
@@ -479,7 +476,7 @@ func (c *SchemaTable) fetchColumnsDefinitions(ctx context.Context, schemaName st
 	}
 
 	if err := idxRS.Err(); err != nil {
-		level.Error(c.logger).Log("msg", "error during iterating over table indexes result set", "schema", schemaName, "table", tableName, "err", err)
+		level.Error(c.logger).Log("msg", "failed to iterate over table indexes result set", "schema", schemaName, "table", tableName, "err", err)
 		return nil, err
 	}
 
@@ -506,7 +503,7 @@ func (c *SchemaTable) fetchColumnsDefinitions(ctx context.Context, schemaName st
 	}
 
 	if err := fkRS.Err(); err != nil {
-		level.Error(c.logger).Log("msg", "error during iterating over foreign keys result set", "schema", schemaName, "table", tableName, "err", err)
+		level.Error(c.logger).Log("msg", "failed to iterate over foreign keys result set", "schema", schemaName, "table", tableName, "err", err)
 		return nil, err
 	}