+++ internal/component/database_observability/mysql/collector/setup_consumers.go
@@ -51,7 +51,7 @@ func NewSetupConsumer(args SetupConsumerArguments) (*SetupConsumers, error) {
 		registry:             args.Registry,
 		setupConsumersMetric: setupConsumerMetric,
 		running:              &atomic.Bool{},
-		logger:               args.Logger,
+		logger:               log.With(args.Logger, "collector", SetupConsumersName),
 		collectInterval:      args.CollectInterval,
 	}, nil
 }
@@ -61,7 +61,7 @@ func (c *SetupConsumers) Name() string {
 }
 
 func (c *SetupConsumers) Start(ctx context.Context) error {
-	level.Debug(c.logger).Log("msg", SetupConsumersName+" collector started")
+	level.Debug(c.logger).Log("msg", "collector started")
 	c.running.Store(true)
 
 	ctx, cancel := context.WithCancel(ctx)
@@ -78,7 +78,7 @@ func (c *SetupConsumers) Start(ctx context.Context) error {
 
 		for {
 			if err := c.getSetupConsumers(c.ctx); err != nil {
-				level.Error(c.logger).Log("msg", "setupConsumers collector error", "err", err)
+				level.Error(c.logger).Log("msg", "collector error", "err", err)
 			}
 
 			select {
@@ -123,7 +123,7 @@ func (c *SetupConsumers) getSetupConsumers(ctx context.Context) error {
 	for rs.Next() {
 		var consumer consumer
 		if err := rs.Scan(&consumer.name, &consumer.enabled); err != nil {
-			return fmt.Errorf("error scanning getSetupConsumers row: %w", err)
+			return fmt.Errorf("failed to scan getSetupConsumers row: %w", err)
 		}
 
 		if consumer.enabled == "YES" {